mezzanine-playground/requirements-humans.txt
django-extensions
mezzanine
pyreadline # Console tab completion
